SpringBoot: Run app with custom environment

+++ Microservices (H2, JPA, WEB, Lombok, Validation, Devtools) - Building Microservice/accounts/src/main/java/com/didan/microservices/accounts/dto/AccountsDto.java
@@ -14,7 +14,7 @@ public class AccountsDto {
     @NotEmpty(message = "AccountNumber can not be a null or empty")
     @Pattern(regexp="(^$|[0-9]{10})",message = "AccountNumber must be 10 digits")
     @Schema(description = "Account Number of account", example = "0123456789")
-	private int accountNumber;
+	private Long accountNumber;
 	
     @NotEmpty(message = "AccountType can not be a null or empty")
     @Schema(description = "Account type of Eazy Bank account", example = "Savings")
