+++ Microservices (H2, JPA, WEB, Lombok, Validation, Devtools) - Building Microservice/accounts/src/main/java/com/didan/microservices/accounts/entity/Accounts.java
@@ -20,7 +20,7 @@ public class Accounts extends BasicClass{
 	
 	@Id
 	@Column(name = "account_number")
-	private int accountNumber;
+	private Long accountNumber;
 	
 	@Column(name = "account_type")
 	private String accountType;
